how to run on a library with cargo test
shows a simple example of how to test a cargo library (src/lib.src). its not quite as straightforward because its not immediately obvious to the user where cargo makes the binary, but if you pass --verbose to cargo test it will show you. you can pass this binary to pprof and generate nice profiling information even if you are building a rust library and dont have a 'main' program.