Use pkg-config to locate libprofiler
There is no guarantee that libprofiler is available on "standard"
locations, which means that simply trying to link to it with -lprofiler
is bound to fail.

For example: a user may have installed gperftools under their home
directory, in which case this crate cannot find it.

To resolve this, use pkg-config to locate the library.  This allows
this crate to learn all compiler and linker flags (in particular, -L)
needed to use the libprofiler library.